CO-2667 Migrate advanced_translation to V11
