+++ advanced_translation/README.rst
@@ -1,32 +1,70 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-    :alt: License: AGPL-3
-
+====================
 Advanced Translation
 ====================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge2| image:: https://img.shields.io/badge/github-CompassionCH%2Fcompassion--modules-lightgray.png?logo=github
+    :target: https://github.com/CompassionCH/compassion-modules/tree/11.0/advanced_translation
+    :alt: CompassionCH/compassion-modules
+
+|badge1| |badge2| 
 
 Use this module to add application terms that can be translated in context
 for male / female and singular / plural.
 
+**Table of contents**
+
+.. contents::
+   :local:
+
 Usage
 =====
 
+You can go in the Technical Settings/Translations/Advanced Translations
 
-Known issues / Roadmap
-======================
-* add support for objects to define itself male/female based on field gender
-* add support for objects to define itself singular/plural based on recordset
-* add support of ${keyword.src} to fetch advanced translations from templates
-* make a view on the terms to be able to edit translations
+- Enter all terms you want to use
+- User method get of ir.advanced.translation in order to fetch translation
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/CompassionCH/compassion-modules/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/CompassionCH/compassion-modules/issues/new?body=module:%20advanced_translation%0Aversion:%2011.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
 
 Credits
 =======
 
+Authors
+~~~~~~~
+
+* Compassion CH
+
 Contributors
-------------
+~~~~~~~~~~~~
 
 * Emanuel Cino <ecino@compassion.ch>
 
-Maintainer
-----------
+Maintainers
+~~~~~~~~~~~
 
-This module is maintained by `Compassion Switzerland <https://www.compassion.ch>`.
+This module is maintained by Compassion Switzerland.
+
+.. image:: https://upload.wikimedia.org/wikipedia/en/8/83/CompassionInternationalLogo.png
+   :alt: Compassion Switzerland
+   :target: https://www.compassion.ch
+
+Compassion Switzerland is a nonprofit organization whose
+mission is to release children from extreme poverty in Jesus name.
+
+This module is part of the `CompassionCH/compassion-modules <https://github.com/CompassionCH/compassion-modules/tree/11.0/advanced_translation>`_ project on GitHub.
