+++ advanced_translation/demo/ir.advanced.translation.csv
@@ -1,0 +1,2 @@
+id,src,lang,group,male_singular,male_plural,female_singular,female_plural
+demo_his,his,en_US,Partner,his,their(m),her,their(f)
