+++ advanced_translation/models/__init__.py
@@ -1,12 +1,3 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Copyright (C) 2016 Compassion CH (http://www.compassion.ch)
-#    Releasing children from poverty in Jesus' name
-#    @author: Emanuel Cino <ecino@compassion.ch>
-#
-#    The licence is in the file __manifest__.py
-#
-##############################################################################
-
 from . import ir_advanced_translation
+from . import translatable_model
+from . import res_partner_test
