+++ advanced_translation/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Emanuel Cino <ecino@compassion.ch>
