+++ advanced_translation/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+Use this module to add application terms that can be translated in context
+for male / female and singular / plural.
