+++ advanced_translation/readme/USAGE.rst
@@ -1,0 +1,4 @@
+You can go in the Technical Settings/Translations/Advanced Translations
+
+- Enter all terms you want to use
+- User method get of ir.advanced.translation in order to fetch translation
