+++ advanced_translation/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_advanced_translation
